fix(nested): integer key-synth casts + friendly nested object over-cardinality + orphan-drop docs
Critical: cast the synthesized nested-input join-key value to the PK
column's own CLR type at the single synthesis site in
SQuiLDataContext.cs (EmitInputFlatten/EmitNode), so smallint (short) and
tinyint (byte) PK/FK columns no longer emit CS1503 (int passed into a
narrower positional ctor arg). int/bigint now go through the same
explicit (harmless) cast for one uniform code path. Synthesizable-type
set is unchanged (int/bigint/smallint/uniqueidentifier; tinyint already
routes to SP0036) - no editor files touched, parity intact. Added
SmallintKeyInputNesting (the RED->GREEN fixture, manually confirmed to
fail with the exact predicted CS1503 before the fix) and
BigintKeyInputNesting (locks the widening boundary) to
SQuiL.Tests/NestedObjects/NestedInputTests.cs.

Important: the nested-output object stitch (EmitNestedReader) now
throws the same friendly "Return object results in more than one
object. Consider using a return table instead." message on a >1-row
object match, instead of the raw LINQ SingleOrDefault
InvalidOperationException, via a new EmitSingleOrFriendly helper used
at both the child-object stitch and object-root assignment sites. 0/1
row behavior is unchanged. Existing EmbeddedObjectChild /
ThreeLevelListNesting / NestedPlusUnrelatedSiblings snapshots updated
to show the new guard.

Important: documented the previously-undocumented orphan-child-drop
behavior (a child row whose key matches no parent row is dropped from
the tree entirely) in SQuiL.Editor.Shared/guide.html and
plugins/squil/skills/squil/SKILL.md; regenerated the 3 per-extension
guide.html copies via tools/GuideRenderer and mirrored the SKILL.md
update to the machine-local copy (hash-identical).

282/282 tests passing (280 + 2 new fixtures).

+++ SQuiL.Tests/NestedObjects/EmbeddedObjectChildInputNesting/NestedInputTests.EmbeddedObjectChildInputNesting#EmbeddedObjectChildInputNestingDataContext.g.verified.cs
@@ -35,14 +35,14 @@ partial class EmbeddedObjectChildInputNestingDataContext : SQuiLBaseDataContext
 		
 		if (request.Order is not null)
 		{
-			var orderKey = __Order.Count + 1;
+			var orderKey = (int)(__Order.Count + 1);
 			__Order.Add(new(
 				orderKey,
 				request.Order.CustomerName));
 			if (request.Order.Shipment is not null)
 			{
 				var shipment = request.Order.Shipment;
-				var shipmentKey = __Shipment.Count + 1;
+				var shipmentKey = (int)(__Shipment.Count + 1);
 				__Shipment.Add(new(
 					shipmentKey,
 					orderKey,
